Add new attributes and methods to user and product models
A new endpoint was added to UserController to fetch users by store id and their roles. `model` and `category` fields have been introduced into `ProductEntity` and `ProductDTO` classes. Similarly, a `phoneNo` field is also added to `UserEntity` and `UserDTO`. The UserRepository and UserService were modified to support the aforementioned changes in the UserController.