Add soft delete functionality and seen status for notifications
Implemented soft delete functionality for ResourceFile and Notification by adding a Boolean field isDeleted. The delete operations are updated to just mark the record as deleted instead of removing it from the database. A "isSeen" field is added to Notification to keep track of whether the notification has been seen by the user. A new method to mark a notification as seen is also added.