Add TransactionHelper class and refactor TransactionService
The commit introduces a new 'TransactionHelper' java class in the transaction/helper package. The helper class aims to streamline the conversion of TransactionEntity to ReturnTransactionDTO in a defined method. A method of the newly created helper class is called in 'TransactionService', leading to the refactoring of 'getTransactionById' method in TransactionService.