Add functionality to fetch top 3 employees by store profit
A new HTTP GET endpoint /top3EmployeesByStoreProfit has been added to the UserController to support fetching of the top 3 employees by their store profit. A corresponding query method was added to UserRepository to find all employees. A service function was added in UserService for the computation of the total profit made by each employee and selecting the top 3 employees. EmployeeProfitDTO was created to hold the information of the employee and their total profit.