Refactor transaction operations and introduce more triggers
The TransactionService has been significantly refactored to introduce more explicit handling for product stock changes occurred during a transaction. In the TransactionController, the previous method for updating a transaction has been commented out for now. The stock update triggers, StockUpdateTriggerUtil and the newly added StockDecreaseTriggerUtil, have been updated to handle stock changes dynamically depending on whether a transaction is for buying or not.